doppelgaenger-websocket/requirements.txt
motor
tornado
dnspython
logzero
python-dotenv
bson